requirements.txt

streamlit
pandas
fpdf
openpyxl
google-api-python-client
google-auth
google-auth-oauthlib
